Move homepage into shared site shell
